requirements.txt
# Core Dependencies
mem0ai==0.1.0
ollama==0.4.0

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
qdrant-client==1.12.0
sqlalchemy==2.0.35
alembic==1.14.0
redis==5.2.0  # Optional for caching

# NLP & ML
transformers==4.46.0
torch==2.5.0  # CPU version, GPU: torch==2.5.0+cu121
sentence-transformers==3.3.0
numpy==1.26.4
scikit-learn==1.5.2

# Text Processing
langchain==0.3.0
langchain-community==0.3.0
tiktoken==0.8.0
beautifulsoup4==4.12.3
markdownify==0.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Frontend (for Streamlit option)
streamlit==1.40.0  # Alternative to React

# Utilities
httpx==0.27.2
aiofiles==24.1.0
python-magic==0.4.27
pandas==2.2.3
openpyxl==3.1.5  # For Excel import/export

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Monitoring (Optional)
prometheus-client==0.21.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0

# Logging
loguru==0.7.2
rich==13.9.0